+++ FireFlight/FireFlight/Networking/APIController.swift
@@ -188,9 +188,9 @@ class APIController {
     
     // MARK: - Post new Address
     
-    func postAddress(label: String, address: String, location: CLLocation, shownFireRadius: Float, completion: @escaping (Error?) -> Void) {
+    func postAddress(label: String, address: String, location: CLLocation, shownFireRadius: Int, completion: @escaping (Error?) -> Void) {
         
-        let newAddress = UserAddress(id: nil, latitude: location.coordinate.latitude, longitude: location.coordinate.longitude, address: address, label: label, radius: Double(shownFireRadius))
+        let newAddress = UserAddress(id: nil, latitude: location.coordinate.latitude, longitude: location.coordinate.longitude, address: address, label: label, radius: shownFireRadius)
         
         let requestURL = Config.baseURL.appendingPathComponent("locations")
         var request = URLRequest(url: requestURL)
@@ -219,9 +219,9 @@ class APIController {
     
     // MARK: - Edit existing address
     
-    func editAddress(id: Int, label: String, address: String, location: CLLocation, shownFireRadius: Float, completion: @escaping (Error?) -> Void) {
+    func editAddress(id: Int, label: String, address: String, location: CLLocation, shownFireRadius: Int, completion: @escaping (Error?) -> Void) {
         
-        let editedAddress = UserAddress(id: id, latitude: location.coordinate.latitude, longitude: location.coordinate.longitude, address: address, label: label, radius: Double(shownFireRadius))
+        let editedAddress = UserAddress(id: id, latitude: location.coordinate.latitude, longitude: location.coordinate.longitude, address: address, label: label, radius: shownFireRadius)
         
         let requestURL = Config.baseURL.appendingPathComponent("locations").appendingPathComponent("\(id)")
         var request = URLRequest(url: requestURL)
